# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is a OpenAPI 2 specification for Mashrooy"
  version: "1.0.0"
  title: "OpenAPI Mashrooy"
host: "localhost:3006"
basePath: "/Mashrooy"
...
schemes:
- "http"
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /group:
    get:
      tags:
      - "Group"
      summary: "list groups of permessions"
      operationId: "listGroup"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "x-access-token"
        in: "header"
        description: "user token to check if user authenticated"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          headers:
            X-Rate-Limit:
              type: "integer"
              format: "int32"
              description: "calls per hour allowed by the employee"
            X-Expires-After:
              type: "string"
              format: "date-time"
              description: "date in UTC when token expires"
          schema:
            type: "string"
        400:
          description: "Invalid username/password supplied"
      x-swagger-router-controller: "Group"
    post:
      tags:
      - "Group"
      summary: "ceate group"
      description: "This can only be done by the logged in Employee."
      operationId: "createGroup"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Created Group array"
        required: true
        schema:
          type: "array"
          items:
            $ref: "#/definitions/Group"
      - name: "x-access-token"
        in: "header"
        description: "user token to check if user authenticated"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Group"
        400:
          description: "Validation Error"
          schema:
            type: "string"
        403:
          description: "Forbidden"
          schema:
            type: "string"
        404:
          description: "Record Not Found"
          schema:
            type: "string"
        409:
          description: "Conflict record"
          schema:
            type: "string"
        500:
          description: "Internal Server Error"
          schema:
            type: "string"
      x-swagger-router-controller: "Group"
  /group/{id}:
    get:
      tags:
      - "Group"
      summary: "Get employee by Id"
      operationId: "getGroupById"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The name that needs to be fetched. Use user1 for testing."
        required: true
        type: "string"
      - name: "x-access-token"
        in: "header"
        description: "user token to check if user authenticated"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
        400:
          description: "Invalid username supplied"
        404:
          description: "User not found"
      x-swagger-router-controller: "Group"
    put:
      tags:
      - "Group"
      summary: "Updated Group"
      description: "This can only be done by the logged in employee."
      operationId: "updateGroupById"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id"
        in: "path"
        description: "The id of the group to update"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Updated employee object"
        required: true
        schema:
          $ref: "#/definitions/Group"
      - name: "x-access-token"
        in: "header"
        description: "user token to check if user authenticated"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Group"
        400:
          description: "Validation Error"
          schema:
            type: "string"
        403:
          description: "Forbidden"
          schema:
            type: "string"
        404:
          description: "Record Not Found"
          schema:
            type: "string"
        409:
          description: "Conflict record"
          schema:
            type: "string"
        500:
          description: "Internal Server Error"
          schema:
            type: "string"
      x-swagger-router-controller: "Group"
    delete:
      tags:
      - "Group"
      summary: "Delete group"
      description: "This can only be done by the logged in employee."
      operationId: "deleteGroup"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The Group that needs to be deleted"
        required: true
        type: "string"
      - name: "x-access-token"
        in: "header"
        description: "user token to check if user authenticated"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
        400:
          description: "Invalid ID supplied"
          schema:
            type: "string"
        403:
          description: "Forbidden"
          schema:
            type: "string"
        404:
          description: "Record Not Found"
          schema:
            type: "string"
        500:
          description: "Internal Server Error"
          schema:
            type: "string"
      x-swagger-router-controller: "Group"
...
securityDefinitions:
  loanstore_auth:
    type: "oauth2"
    authorizationUrl: "https://tasaheel.swagger.io/oauth/authorize"
    flow: "implicit"
    scopes:
      write:loans: "modify loans in your account"
      read:loans: "read your loans"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  Group:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        $ref: "#/definitions/Group_name"
      permissions:
        type: "string"
    example:
      permissions: "permissions"
      name:
        ar: "ar"
        en: "en"
      id: 0
...
  Group_name:
    type: "object"
    properties:
      ar:
        type: "string"
      en:
        type: "string"
    example:
      ar: "ar"
      en: "en"
...